third_party/WebKit/LayoutTests/http/tests/bluetooth/https/requestDevice/cross-origin-iframe.html

<!DOCTYPE html>
<script src="/js-test-resources/testharness.js"></script>
<script src="/js-test-resources/testharnessreport.js"></script>
<script src="/js-test-resources/bluetooth/bluetooth-helpers.js"></script>
<body>
  <script>
  "use strict";
  async_test(test => {
    window.onmessage = messageEvent => test.step(() => {
      if (messageEvent.data === 'Ready') {
        let iframe = document.querySelector('iframe');
        callWithKeyDown(() => {
          iframe.contentWindow.postMessage('Go', '*');
        });
      } else {
        assert_equals(messageEvent.data, 'SecurityError: Access to the Web ' +
                                         'Bluetooth API is disallowed by ' +
                                         'feature policy.');
        test.done();
      }
    });
    setBluetoothFakeAdapter('HeartRateAdapter')
      .then(() => {
        let iframe = document.createElement('iframe');
        iframe.src = 'https://localhost:8443/js-test-resources/bluetooth/requestDevice-in-iframe.html';
        document.body.appendChild(iframe);
      });
  }, 'Request device from a unique origin. Should reject with SecurityError.');
  </script>
</body>
